packages/podcast-generator/requirements.txt
firebase-admin==6.2.0
requests==2.31.0
pydub==0.25.1
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
